Add status reporting watcher
